chore(folder): rename files to index
