+++ src/components/Button/test.tsx
@@ -1,7 +1,7 @@
-import Button from './Button'
-
 import React from 'react'
 import { screen, render } from '@testing-library/react'
+
+import Button from './'
 
 describe('Button', () => {
   it('deve renderizar um botão com o texto passado', () => {
