+++ src/components/Heading/test.tsx
@@ -1,7 +1,7 @@
-import Heading from './Heading'
-
 import React from 'react'
 import { screen, render } from '@testing-library/react'
+
+import Heading from './'
 
 describe('Heading', () => {
   it('shows Heading information', () => {
