+++ src/pages/index.mdx
@@ -1,4 +1,4 @@
-import Button from '../components/Button/Button'
+import Button from '../components/Button'
 
 # Olá mundo!
